Added functionality to the floating button
