+++ app/src/main/java/com/apptec/registrateapp/mainactivity/MainViewModel.java
@@ -245,5 +245,4 @@ public class MainViewModel extends AndroidViewModel {
     }
 
 
-
 }
