+++ app/src/main/java/com/apptec/registrateapp/mainactivity/fpermission/ui/PermissionFragment.java
@@ -1,9 +1,11 @@
 package com.apptec.registrateapp.mainactivity.fpermission.ui;
 
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ListView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -14,6 +16,7 @@ import androidx.lifecycle.ViewModelProviders;
 import com.apptec.registrateapp.R;
 import com.apptec.registrateapp.mainactivity.MainViewModel;
 import com.apptec.registrateapp.models.PermissionModel;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.util.List;
 
@@ -21,11 +24,16 @@ public class PermissionFragment extends Fragment {
     /**
      * PermissionFragment
      */
+    private static final String TAG = "PermissionFragment";
+
 
     // Instance of ViewModel
     private MainViewModel mainViewModel;
 
     // Ui elements
+    FloatingActionButton floatingActionButton;
+    ListView permissionsListView;
+    PermissionListAdapter permissionListAdapter;
 
 
     @Override
@@ -39,14 +47,32 @@ public class PermissionFragment extends Fragment {
                              ViewGroup container, Bundle savedInstanceState) {
 
         View view = inflater.inflate(R.layout.fragment_permission, container, false);
-
+        floatingActionButton = view.findViewById(R.id.add_permission_floating_button);
         // TODO: Observe the mPermissionList
         mainViewModel.getPermissionsList().observe(this, new Observer<List<PermissionModel>>() {
             @Override
             public void onChanged(List<PermissionModel> permissionModels) {
                 // TODO
+
             }
         });
+
+
+        floatingActionButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                /**
+                 * Open the dialog fragment to add a permission
+                 */
+                Log.d(TAG, "Open dialog");
+
+                DialogPermission dialogPermission = new DialogPermission();
+                dialogPermission.show(getFragmentManager(), TAG);
+
+
+            }
+        });
+
 
         return view;
     }
